requirements.txt
click==8.0.3
hydra-core==1.1.1
matplotlib==3.4.3
numpy==1.20.3
pytest==6.2.4
python-dotenv==0.19.2
setuptools==58.0.4
torch==1.10.0
torchvision==0.11.1
